AuthConfig metrics extracted to separate file
